% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx.mixEM}
\alias{as.fmx.mixEM}
\title{Convert \code{mixEM} Objects to \linkS4class{fmx} Objects}
\usage{
\method{as.fmx}{mixEM}(x, data = x[["x"]], ...)
}
\arguments{
\item{x}{\code{mixEM} object}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{..}
}
\value{
Function \code{\link[=as.fmx.mixEM]{as.fmx.mixEM()}} returns an \linkS4class{fmx} object.
}
\description{
..
}
\note{
\link[mixtools]{plot.mixEM} not plot \link[mixtools]{gammamixEM} returns, as of 2022-09-19.
}
\examples{
library(mixtools)
(x = as.fmx(normalmixEM(faithful$waiting, k = 2)))

}
