% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmxS.R
\name{autoplot.fmxS}
\alias{autoplot.fmxS}
\title{Plot \linkS4class{fmxS} Object}
\usage{
\method{autoplot}{fmxS}(
  object,
  type = c("density", "distribution"),
  xlim = if (length(data)) range.default(data) else range.default(lapply(dots, FUN =
    qfmx, p = c(0.01, 0.99))),
  ...
)
}
\arguments{
\item{object}{\linkS4class{fmxS} object}

\item{type}{\link[base]{character} scalar, \code{'density'} (default) or \code{'distribution'}}

\item{xlim}{..}

\item{...}{..}
}
\value{
Function \code{\link[=autoplot.fmxS]{autoplot.fmxS()}} returns a \link[ggplot2]{ggplot} figure.
}
\description{
..
}
