% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_init.R
\name{clusterList}
\alias{clusterList}
\title{Clustering Observations for Creation of \linkS4class{fmx} Object}
\usage{
clusterList(x, K, method = c("reassign_tkmeans"), alpha = 0.05, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, one-dimensional observations}

\item{K}{\link[base]{integer} scalar, number of mixture components}

\item{method}{\link[base]{character} scalar, only \code{'reassign_tkmeans'} supported yet}

\item{alpha}{\link[base]{numeric} scalar, proportion of observations to be trimmed in
trimmed k-means algorithm \link[tclust]{tkmeans}}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{clusterList} returns a \link[base]{list} of \link[base]{numeric} \link[base]{vector}s.
}
\description{
..
}
\seealso{
\link[stats]{kmeans} \link[tclust]{tkmeans} \link{reAssign}
}
