% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_constraint.R
\name{constraint_TeX}
\alias{constraint_TeX}
\title{TeX Label of Parameter Constraint(s) of \linkS4class{fmx} Object}
\usage{
constraint_TeX(dist)
}
\arguments{
\item{dist}{\linkS4class{fmx} object}
}
\value{
\link{constraint_TeX} returns a \link[base]{character} scalar
(of TeX expression) of the constraint,
primarily intended for end-users in plots.
}
\description{
Create TeX label of parameter constraint(s) of \linkS4class{fmx} object
}
\examples{
(d0 = fmx('GH', A = c(1,4), g = c(.2,.1), h = c(.05,.1), w = c(1,1)))
constraint_TeX(d0)

(d1 = fmx('GH', A = c(1,4), g = c(.2,0), h = c(0,.1), w = c(1,1)))
constraint_TeX(d1)

(d2 = fmx('GH', A = c(1,4), g = c(.2,0), h = c(.15,.1), w = c(1,1)))
constraint_TeX(d2)

}
