% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_repar.R
\name{dbl2fmx}
\alias{dbl2fmx}
\title{Inverse of \link{fmx2dbl}, for internal use}
\usage{
dbl2fmx(x, K, distname, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, unrestricted parameters}

\item{K}{\link[base]{integer} scalar}

\item{distname}{\link[base]{character} scalar}

\item{...}{additional parameters, not currently used}
}
\value{
\link{dbl2fmx} returns a \link[base]{list} with two elements \code{'pars'} and \code{'w'}
}
\description{
..
}
\details{
Only used in \link{QLMDe} and unexported function \code{qfmx_gr}, not compute intensive
}
