% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx.R
\name{K.fmx}
\alias{K.fmx}
\title{Number of Components in \code{\linkS4class{fmx}} and \code{\linkS4class{fmx_QLMDe}} Object}
\usage{
K.fmx(x)
}
\arguments{
\item{x}{\code{\linkS4class{fmx}} and \code{\linkS4class{fmx_QLMDe}} object.}
}
\value{
An \code{\link[base]{integer}} value indicating the number of components in 
an \code{\linkS4class{fmx}} and.or \code{\linkS4class{fmx_QLMDe}} object.
}
\description{
Obtain the number of components in \code{\linkS4class{fmx}} and \code{\linkS4class{fmx_QLMDe}} object.
}
\details{
For user convenience
}
\examples{

(d2 = fmx('GH', A = c(1,6), B = 2, g = c(0,.3), h = c(.2,0), w = c(1,2)))
K.fmx(d2)

}
