% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_dist.R
\name{KL_dist}
\alias{KL_dist}
\title{Kullback-Leibler Divergence for Distribution Estimates}
\usage{
KL_dist(x, base, ...)
}
\arguments{
\item{x}{an R object of distribution estimates}

\item{base}{see \code{\link[LaplacesDemon]{KLD}}}

\item{...}{additional parameters, currently not in use}
}
\value{
\code{\link{KL_dist}} returns a \code{\link[base]{list}}, which is returned from \code{\link[LaplacesDemon]{KLD}}.
}
\description{
Calculate the Kullback-Leibler divergence for distribution estimates, via \code{\link[LaplacesDemon]{KLD}}.
}
\seealso{
\code{\link[LaplacesDemon]{KLD}}.
}
