% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmx_init.R
\name{klist}
\alias{klist}
\title{Data Clusters by (modified) \eqn{k}-Means}
\usage{
klist(x, K, method = c("reassign_tkmeans"), alpha = 0.05, ...)
}
\arguments{
\item{x}{\link[base]{numeric} \link[base]{vector}, one-dimensional observations}

\item{K}{\link[base]{integer} scalar, number of clusters}

\item{method}{\link[base]{character} scalar,
only \code{'reassign_tkmeans'} (default) supported yet}

\item{alpha}{\link[base]{numeric} scalar, proportion of observations to be trimmed in
trimmed \eqn{k}-means algorithm \link[tclust]{tkmeans}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=klist]{klist()}} returns a \link[base]{list} of \link[base]{numeric} \link[base]{vector}s.
}
\description{
To create a \link[base]{list} of observations based on (modified) \eqn{k}-means algorithm.
}
\seealso{
\link{reAssign.tkmeans}
\link[stats]{kmeans}
}
\keyword{internal}
