% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\title{Kolmogorov-Smirnov Tests for Distribution Estimates}
\usage{
ks_test(x, data, ...)
}
\arguments{
\item{x}{an R object of distribution estimates}

\item{data}{\link[base]{double} vector, the actual observations}

\item{...}{additional parameters of \link[stats]{ks.test}}
}
\value{
\link{ks_test} returns an \link[stats:ks.test]{htest} object,
in which the element \code{$statistic} is the Kolmogorov–Smirnov distance.
}
\description{
Perform the Kolmogorov-Smirnov tests for various distribution estimates.
}
\seealso{
\link[stats]{ks.test}
}
