% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitdist_ext.R
\name{logLik.fitdist}
\alias{logLik.fitdist}
\title{Log-Likelihood of \link[fitdistrplus]{fitdist} Object}
\usage{
\method{logLik}{fitdist}(object, ...)
}
\arguments{
\item{object}{\link[fitdistrplus]{fitdist} object}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{logLik.fitdist} returns a \link[stats]{logLik} object
}
\description{
..
}
\details{
Output of \link[fitdistrplus]{fitdist} has element \code{$loglik}
(as well as \code{$aic} and \code{$bic}),
but it is simply a \link[base]{numeric} scalar.
\code{fitdistrplus:::logLik.fitdist} simply returns this element.

\link{logLik.fitdist} returns a \link[stats]{logLik} object, which
could be further used by \link[stats]{AIC} and \link[stats]{BIC}.

(I have written to the authors)
}
\seealso{
\link[fitdistrplus]{fitdist}
}
