% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsmsn_dpqr.R
\name{mixsmsn_skewNormal}
\alias{mixsmsn_skewNormal}
\alias{dSN}
\alias{pSN}
\title{Skew Normal as in \CRANpkg{mixsmsn}}
\usage{
dSN(x, mean = 0, sd = 1, shape = 0, log = FALSE)

pSN(q, mean = 0, sd = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{..}

\item{mean}{..}

\item{sd}{..}

\item{shape}{..}

\item{log, log.p}{..}

\item{lower.tail}{..}
}
\value{
\link{dSN} returns a \link[base]{numeric} \link[base]{vector} of the same length as \code{x},
or a \link[base]{numeric} \link[base]{matrix} of the same dimension as \code{x}.
}
\description{
..
}
\seealso{
\code{mixsmsn:::dSN} \link[sn]{dsn}
}
