% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsmsn_dpqr.R
\name{mixsmsn_skewT}
\alias{mixsmsn_skewT}
\alias{dST}
\alias{pST}
\title{Skew \eqn{t} as in \CRANpkg{mixsmsn}}
\usage{
dST(x, mean = 0, sd = 1, shape = 0, nu = 4, log = FALSE)

pST(q, mean = 0, sd = 1, shape = 0, nu = 4, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{..}

\item{mean}{..}

\item{sd}{..}

\item{shape}{..}

\item{nu}{..}

\item{log, log.p}{..}

\item{lower.tail}{..}
}
\value{
\link{dST} returns a \link[base]{numeric} \link[base]{vector} of the same length as \code{x},
or a \link[base]{numeric} \link[base]{matrix} of the same dimension as \code{x}.
}
\description{
..
}
\seealso{
\code{mixsmsn:::dt.ls} \link[sn]{dst}
}
