% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_fmx.R
\name{nobs.fmx}
\alias{nobs.fmx}
\title{Number of Observations in \linkS4class{fmx} Object}
\usage{
\method{nobs}{fmx}(object, ...)
}
\arguments{
\item{object}{\linkS4class{fmx} object}

\item{...}{place holder for S3 naming convention}
}
\value{
\link{nobs.fmx} returns an \link[base]{integer} scalar.
}
\description{
..
}
\details{
\link{nobs.fmx} returns the sample size of
the observations used in \link{QLMDe} estimation, or \code{integer(0)} for distribution-only
\linkS4class{fmx} object
}
\seealso{
\link[stats]{nobs}
}
