% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment_vanilla.R
\name{show,moment-method}
\alias{show,moment-method}
\title{Show \linkS4class{moment}}
\usage{
\S4method{show}{moment}(object)
}
\arguments{
\item{object}{\linkS4class{moment}}
}
\value{
The \link[methods]{show} method for \linkS4class{moment} object
does not have a returned value.
}
\description{
..
}
