% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0S4GenericDoc.R, R/fmx.R
\name{show_S4}
\alias{show_S4}
\alias{show,fmx-method}
\title{Documentation for S4 Generic \code{\link[methods]{show}}}
\usage{
\S4method{show}{fmx}(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
An invisible \code{NULL} is returned, following the behavior of \code{\link[methods]{show}}.
}
\description{
Automatically display objects from an S4 class (a class defined by a call to \code{\link[methods]{setClass}}).
}
\seealso{
\code{\link[methods]{show}}
}
