% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixsmsn_ext.R
\name{sort_mixsmsn}
\alias{sort_mixsmsn}
\alias{sort.Skew.normal}
\alias{sort.Normal}
\alias{sort.Skew.t}
\alias{sort.t}
\title{Sort Objects from \pkg{mixsmsn} by Location Parameters}
\usage{
\method{sort}{Skew.normal}(x, decreasing = FALSE, ...)

\method{sort}{Normal}(x, decreasing = FALSE, ...)

\method{sort}{Skew.t}(x, decreasing = FALSE, ...)

\method{sort}{t}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{\code{Normal}, \code{Skew.normal}, \code{Skew.t} object}

\item{decreasing}{\link[base]{logical} scalar. Should the sort the location parameter
be increasing (\code{FALSE}, default) or decreasing (\code{TRUE})?}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{sort.Normal} returns a \code{Normal} object.

\link{sort.Skew.normal} returns a \code{Skew.normal} object.

\link{sort.Skew.t} returns a \code{Skew.t} object.
}
\description{
To sort an object returned from package \pkg{mixsmsn} by its location parameters
}
\details{
\link[mixsmsn]{smsn.mix} does \strong{not} order the location parameter
}
\seealso{
\link[base]{sort}
}
