% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allequal.R
\name{allequal}
\alias{allequal}
\alias{unique_allequal}
\alias{duplicated_allequal}
\title{Determine Nearly-Equal Elements}
\usage{
unique_allequal(x, ...)

duplicated_allequal(x, ...)
}
\arguments{
\item{x}{\link[base]{double} \link[base]{vector}}

\item{...}{additional parameters of function \link{outer_allequal}}
}
\value{
Function \link{duplicated_allequal} returns a \link[base]{logical} \link[base]{vector} of the same length as the input vector,
indicating whether each element is nearly-equal to any of the previous elements.

Function \link{unique_allequal} returns the non-nearly-equal elements in the input vector.
}
\description{
Determine nearly-equal elements and extract non-nearly-equal elements in a \link[base]{double} \link[base]{vector}.
}
\examples{
x = c(.3, 1-.7, 0, .Machine$double.eps)
unique.default(x) # not desired
unique_allequal(x) # desired

}
\seealso{
\link[base]{duplicated.default} \link[base]{unique.default}
}
\keyword{internal}
