% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fmx.R
\name{as.fmx}
\alias{as.fmx}
\title{Turn Various Objects to \linkS4class{fmx}}
\usage{
as.fmx(x, data, ...)
}
\arguments{
\item{x}{an R object}

\item{data}{\link[base]{numeric} \link[base]{vector}}

\item{...}{..}
}
\value{
\link{as.fmx} returns an \linkS4class{fmx} object.
}
\description{
Turn various objects that are created in other packages
to \linkS4class{fmx} class
}
\details{
In order to take advantage of all methods for \linkS4class{fmx} objects
}
\seealso{
\link{as.fmx.fitdist} \link{as.fmx.mixEM}
}
