% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM_test.R
\name{CvM_test}
\alias{CvM_test}
\title{Cramer-Von Mises Test of Goodness-of-Fit for Distribution Estimates}
\usage{
CvM_test(x, nullname, ...)
}
\arguments{
\item{x}{an R object of distribution estimates}

\item{nullname, ...}{additional parameters of \code{\link[goftest]{cvm.test}}}
}
\value{
\code{\link[goftest]{cvm.test}}

\code{\link{CvM_test}} returns the \code{'htest'} object (returned from \code{\link[goftest]{cvm.test}}),
in which the element \code{$statistic} is the Cramer-Von Mises quadratic distance.
}
\description{
Perform the Cramer-Von Mises test of goodness-of-fit for distribution estimates, via \code{\link[goftest]{cvm.test}}.
}
\details{
Note that we are currently not using the discrete version \code{\link[dgof]{cvm.test}}.
}
\seealso{
\code{\link[goftest]{cvm.test}}
}
