% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLMDe.R
\docType{class}
\name{fmx_QLMDe-class}
\alias{fmx_QLMDe-class}
\title{Specification of \linkS4class{fmx_QLMDe} Class}
\description{
Quantile least Mahalanobis distance estimates of one-dimensional finite mixture distribution.
The \linkS4class{fmx_QLMDe} object contains (i.e., inherits from) the \linkS4class{fmx} object.
}
\section{Slots}{

\describe{
\item{\code{data}}{\link[base]{numeric} vector, the one-dimensional observations}

\item{\code{data.name}}{\link[base]{character} scalar, a human-friendly name of observations}

\item{\code{epdf}}{empirical probability density \link[base]{function} fitted by \link[stats]{approxfun}}

\item{\code{quantile_vv}}{variance-covariance \link[base]{matrix} of selected quantiles (based on the selected probabilities stored in slot \code{@p})}

\item{\code{vcov}}{variance-covariance \link[base]{matrix} of the internal (i.e., unconstrained) estimates}

\item{\code{init}}{\linkS4class{fmx} object, the initial values to be sent to \link[stats]{optim}}

\item{\code{p}}{\link[base]{numeric} vectors of probabilities, where the distance between the empirical and true quantiles are measured}

\item{\code{optim}}{a \link[base]{list} returned from \link[stats]{optim}}
}}

