% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_dist.R
\name{KL_dist}
\alias{KL_dist}
\title{Kullback-Leibler Divergence for Distribution Estimates}
\usage{
KL_dist(x, base, ...)
}
\arguments{
\item{x}{an R object of distribution estimates}

\item{base}{see \link[LaplacesDemon]{KLD}}

\item{...}{additional parameters, currently not in use}
}
\value{
\link{KL_dist} returns a \link[base]{list}, which is returned from \link[LaplacesDemon]{KLD} function.
}
\description{
Calculate the Kullback-Leibler divergence for distribution estimates.
}
\seealso{
\link[LaplacesDemon]{KLD}
}
