% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_test.R
\name{ks_test}
\alias{ks_test}
\title{Kolmogorov-Smirnov Tests for Distribution Estimates}
\usage{
ks_test(x, ...)
}
\arguments{
\item{x}{an R object of distribution estimates}

\item{...}{additional parameters of \code{\link[stats]{ks.test}}}
}
\value{
\code{\link{ks_test}} returns the \code{'htest'} object (returned from \code{\link[stats]{ks.test}}),
in which the element \code{$statistic} is the Kolmogorov–Smirnov distance.
}
\description{
Perform the Kolmogorov-Smirnov tests for various distribution estimates, via \code{\link[stats]{ks.test}}.
}
\seealso{
\code{\link[stats]{ks.test}}
}
