% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikRatio.R
\name{LikRatio}
\alias{LikRatio}
\title{Likelihood Ratio Test for General Models}
\usage{
LikRatio(models, type = c("plain", "vuong"), compare = c("seq", "first"), ...)
}
\arguments{
\item{models}{a \code{'list'} of regression models, or a \code{'list'} of \code{'logLik'} objects.}

\item{type}{\code{'character'} value, \code{'plain'} (default) for ordinary likelihood ratio test;
\code{'vuong'} for Vuong's closeness test for non-nested models}

\item{compare}{type of comparison between the models, sequentially (\code{'seq'}) or 
all models versus the first model (\code{'first'})}

\item{...}{potential arguments (currently not in use)}
}
\value{
\code{\link{LikRatio}} returns an ANOVA table of class \code{'anova'} for likelihood ratios test, or class \code{'vuong'} for Vuong's test.
}
\description{
...
}
\details{
Inspired by \code{\link[lmtest]{lrtest.default}}.
}
\examples{
# no examples for now

}
