% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0Int_base_stats.R
\name{mahalanobis_int}
\alias{mahalanobis_int}
\title{A Simpler and Faster Mahalanobis Distance}
\usage{
mahalanobis_int(x, center, invcov)
}
\arguments{
\item{x}{\code{\link[base]{numeric}} vector}

\item{center}{\code{\link[base]{numeric}} vector}

\item{invcov}{the \emph{inverted} covariance \code{\link[base]{matrix}}}
}
\value{
A \code{\link[base]{numeric}} scalar.
}
\description{
Mahalanobis distance, simpler and faster than \code{\link[stats]{mahalanobis}}.
}
\seealso{
\code{\link[stats]{mahalanobis}}
}
