% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{QuantileGH-package}
\alias{QuantileGH}
\alias{QuantileGH-package}
\title{Quantile Least Mahalanobis Distance Estimator for Tukey \eqn{g}-&-\eqn{h} Mixture}
\description{
Tools for simulating and fitting finite mixtures of the 4-parameter Tukey \eqn{g}-&-\eqn{h} distributions.
Tukey \eqn{g}-&-\eqn{h} mixture is highly flexible to model multimodal distributions with variable degree of skewness and kurtosis in the components.
The Quantile Least Mahalanobis Distance estimator \link{QLMDe} is used for estimating parameters of the finite Tukey \eqn{g}-&-\eqn{h} mixtures.
\link{QLMDe} is an indirect estimator that minimizes the Mahalanobis distance between the sample and model-based quantiles.
A backward-forward stepwise model selection algorithm is provided to find
\itemize{
\item {a parsimonious Tukey \eqn{g}-&-\eqn{h} mixture model, conditional on a given number-of-components; and}
\item {the optimal number of components within the user-specified range.}
}
}
\examples{
# see ?QLMDe

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

Authors:
\itemize{
  \item Inna Chervoneva \email{Inna.Chervoneva@jefferson.edu} (\href{https://orcid.org/0000-0002-9104-4505}{ORCID})
}

}
